{"schema":"decodex/commit/1","summary":"Move scroll sample registration","authority":"manual"}
