+++ packages/rsnap-overlay/src/scroll_capture/downward_resolution.rs
@@ -1,3 +1,5 @@
+use std::ops::RangeInclusive;
+
 use color_eyre::eyre::Result;
 
 use crate::scroll_capture::{
@@ -12,7 +14,7 @@ use crate::scroll_capture::{
 	LOCAL_DOWNWARD_SEARCH_MOTION_TOLERANCE_ROWS, MotionObservation, OverlapSearchConfig,
 	OverlapSearchRange, PREVIEW_ONLY_LOCAL_NEAR_CONTINUITY_ROWS,
 	PREVIEW_ONLY_LOCAL_RECOVERY_MAX_MOTION_ROWS, PREVIEW_ONLY_LOCAL_RECOVERY_MAX_TOLERANCE_ROWS,
-	RangeInclusive, RgbaImage, ScrollDirection, ScrollObserveOutcome, ScrollSession,
+	RgbaImage, ScrollDirection, ScrollObserveOutcome, ScrollSession,
 	TRANSIENT_BURST_UNDERCONSUMED_HINT_MIN_ROWS, UNDERCONSUMED_OBSERVED_BURST_RECOVERY_GAP_ROWS,
 	eyre,
 };
